migrate to new helper file so task generate does not keep wiping off
