+++ internal/graphapi/internalpolicyextended.resolvers.go
@@ -65,42 +65,3 @@ func (r *internalPolicyResolver) LiveExternalContents(ctx context.Context, obj *
 
 	return &export.HTML, nil
 }
-
-// !!! WARNING !!!
-// The code below was going to be deleted when updating resolvers. It has been copied here so you have
-// one last chance to move it out of harms way if you want. There are two reasons this happens:
-//  - When renaming or deleting a resolver the old code will be put in here. You can safely delete
-//    it when you're done.
-//  - You have helper methods in this file. Move them out to keep these resolver files clean.
-/*
-	func (r *internalPolicyResolver) findPrimaryDriveIntegration(ctx context.Context, ownerID string) (*generated.Integration, error) {
-	integrations, err := withTransactionalMutation(ctx).Integration.Query().
-		Where(
-			integration.OwnerID(ownerID),
-			integration.DefinitionID(googledrive.DefinitionID()),
-			integration.StatusEQ(enums.IntegrationStatusConnected),
-		).
-		All(ctx)
-	if err != nil {
-		logx.FromContext(ctx).Error().Err(err).Msg("failed to query drive integrations")
-		return nil, err
-	}
-
-	for _, integ := range integrations {
-		var input googledrive.UserInput
-		if err := jsonx.UnmarshalIfPresent(integ.Config.ClientConfig, &input); err != nil {
-			continue
-		}
-
-		if input.Primary {
-			return integ, nil
-		}
-	}
-
-	if len(integrations) > 0 {
-		return integrations[0], nil
-	}
-
-	return nil, nil
-}
-*/
